move state and state diff to folder state
