+++ crates/starknet/src/traits.rs
@@ -4,7 +4,7 @@ use starknet_types::contract_storage_key::ContractStorageKey;
 use starknet_types::felt::{Balance, ClassHash, Felt};
 
 use crate::error::Result;
-use crate::state_diff::StateDiff;
+use crate::state::state_diff::StateDiff;
 
 /// This trait should be implemented by structures that internally have collections and each element
 /// could be found by a hash
